requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.60.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mysql-connector-python==9.5.0
numpy==2.3.4
opencv-python==4.13.0.92
packaging==25.0
pandas==2.3.3
pillow==12.1.1
pillow-heif==0.13.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
soupsieve==2.8
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.7
